CHANGE: Mover smooth movement
I decided against making a separate SmoothMover class because I couldn't
think of an instance in which I would use a non-smooth Mover. So I
improved the smooth movement. Before this change, Mover would store the
delta value that hadn't been used in its movement every tick, and try to
use it the next tick, and save the unused delta, and continue that
pattern. I don't know what I was thinking. I scrapped that system, and
just changed the coordinates to doubles, so that a Mover's location can
be .002, for example, and delta can be .3, and the mover will move
steadily at .3 pixels per tick, without all the extra variable and logic
crap. Since I changed the coord data types, I had to override Entity's
get and set methods as well.

I also made the class abstract, so I removed the tick method and the
constructor, since they are unnecessary and won't be called anyway.