OPT: Render speed
The Index and Clip variables used to be initialized to the default at
their declaration, but since the if statements afterwards have to be run
anyway, I decided to put the default initializations in there as else
clauses. I honestly don't think this will have a real effect on render
speed, but it can't hurt and could save a little bit when rendering
objects at the edge of the screen.

Also added more comments.