ADD Err class to print errors to console
Objects can use Err to print standardized error messages. Each message
has an ID number, which is used when calling it. This will make it
easier to standardize error messages in the future, especially to make
changes to every error message in the program. If I want to change the
email contact, I don't want to change every single reference to it
individually, for example.