ADD keyboard support v.0.02_2
KeyInput implements KeyListener to take input from the keyboard. It is
added to world so that its tick method is called, but that may change in
future versions.

KeyInput has an array of booleans. When it recieves a KeyEvent, it sets
the array location of the KeyCode to true if it's pressed, false if it's
not. There are separate booleans corresponding to each control. At
tick(), the corresponding slots in the keys array are checked, and the
individual control keys set accordingly. There will probably be a
simpler system for this in the future.