CHANGE KeyInput implementation
KeyInput used to be plugged into the tick cycle. I felt this was wrong,
since the tick cycle is intended to update game objects, so I modified
it to only store and return data instead of processing it regularly as
it was. And now it will only check key values when they are requested,
instead of constantly checking every variable it owns, so it should be
more efficient as well. And on top of all that, this change allows for
easy implementation of user preference options down the road.

I did this by replacing the single boolean variables it had with int
arrays. The arrays hold keyCode values, meaning they can be used to go
to the correct spot in the 'keys' array. This will mean that there can
be practically unlimited numbers of keys that all do the same thing, and
also that the same key can do multiple different things, allowing for
lots of user customization.