REMOVE LightweightImage getPixelCopy()
LightweightImage's getPixelCopy method is removed here because it is
simply unnecessary. It would be easy to call getPixels().clone() if you
need a cloned array. As it is, it opens the door for accidental array
copying, possibly leading to hanging garbage. Normally I wouldn't mind
but the idea of LightweightImage is to be as lightweight as is
practically possible.