ADD player, CHANGE KeyInput stuff to static
It's about time I add a player object. It can move within the world
bounds when the up down left or right keys are pressed. I'll have to add
proper collision detection later.

KeyInput methods and variables are static because I want any object
anywhere to be able to tell if a key is pressed at any time, without
needing to use a variable. Arguably this could be made better by having
World or a Control class that tells everybody that needs to know when a
key is pressed, but that's another list for things to register
themselves with and I would like to avoid that as much as is practical.