REMOVE delta, FIX Mover overrides, MOVE world var
delta is removed from Mover because, while I may not have been able to
think of it before, there are lots of times when I don't want to use it.
Case in point: Player. So delta and its methods are gone.

setPos() and setExactPos() methods called in Movers will now set both
the exact position and the int position. This is very important because
previously, if you forgot that you had to use setExactPos, you could get
something that never moved at all. That's the kind of annoyance no one
wants to deal with, so now there are smarter overrides that will use
whatever type you give them. To further secure the Entity-Mover
connection, the x and y variables are now private and cannot be accessed
without a method call.

Mover's world variable is now in Entity instead, because it's likely
that any non-mover entities in the future will want to have world access
too.

Also commented Player