Add screenshots
Screenshots are saved to the screenshots folder. The default key binding
is F1, but this can be changed using settings.txt. Screen 1 demonstrates
the date/time naming format. A screenshot can be taken as many times as
you want: There is no time delay, but the screenshot key must be pressed
and released each time.

I decided to put the screenshot methods into Render because they didn't
seem deserving of their own class, and Render seems the most appropriate
place. It already has all the pixels and everything, it would be awkward
to put it anywhere else.

The Err class has been adjusted accordingly. Also refactored some code
in KeyInput and Game.