Add EntityHolder interface
Entity had a problem where it was dependent on the existence of a World
object. That is poor practice, because World could become very
complicated in the future, and Entity would still depend on it. This is
a problem as it severely limits reuse of Entities for other purposes.
They would always have to have a World to reference.

EntityHolder solves this problem by providing a standardized interface
for using Entity objects. All references to World in the Entity class
have been changed to EntityHolder. Now anything can use an Entity just
by implementing EntityHolder.