Refactor game
Game doesn't contain the main method, the Thread information, or the
code to set up the JFrame. That is all incased in a Start class that
takes care of the game. A Game object can be used in any sort of
Container without needing to modify the class itself. This is useful in,
say, a world editor where Game might be just one object among several
different containers used to edit it.