Implement animation
The new LightweightAnimation class is a sibling of LightweightImage.
Both are descendants of GimbalImage. I was originally going to make
LightweightAnimation extend LightweightImage, but this led to a lot of
unused LightweightImage code that just got in the way.

Methods that used to deal with LightweightImages now mostly deal with
GimbalImages.

Also git ignores .gmbl files. I don't want to commit my quicksaves.