Refactor Entity and GimbalImage, Pre-Alpha.0.05
GimbalImage is an interface instead of an abstract class. This is more
elegant since LightweightImage and LightweightAnimation didn't inherit
anything from it except a method to copy an int[]. Everything using
GimbalImage still works the same.

LightweightImage and LightweightAnimation use BufferedImage's getRGB()
method to copy the pixels over instead of using the native method.
That's really the way it should be done anyway.

LightweightAnimation's setImage() method is restructured to make it only
set the width once, after all the frames have been generated. If there's
an error generating the frames, width won't be able to be set, but I
figure this is a non-issue because if there's an error generating the
frames, it's because of the JRE or the file is saved incorrectly. The
game won't work anyway in that case.

Everything relating to location in Entity is abstract. Mover can handle
it alone, in fact Entity's pre-existing variables and methods were only
getting in the way. Movers had both a double and an int for their
coordinates, and that's ridiculous. Any future Entity that doesn't use
doubles like Mover does can do the same as Mover but with ints.

I also commented a bunch of stuff, hooray!