Add basic collision detection, circular movement
Pre-Alpha.0.06

An object implementing Physical can get another Physical object's
rectangle. That's really all there is to it. Later on there will
probably be more features and the Physical interface should help with
that.

Mobs are Physical. They have a hitbox (just a Rectangle) and the ability
to correct for out of bounds.

Baddie moves around randomly. Game has a new static Random that it uses
to generate its numbers. If it used its own Random, then all the
instances would do the same thing since they all get their tick()
methods called in the same millisecond.

KeyInput is in the io package instead of having an input package.

Player and Baddie both use trig to movethe same distance no matter the
angle. Player's is modified because with keyboard input there are
booleans to work with instead of angle values.

Game holds the value of the universal isometric ratio: 0.5. If I ever
want to change the ratio, I just have to change that value and all the
movement will be covered.

Git ignores screenshots/. Deleted the only screenshot so it stops
following it.

player.png wasn't used. It has been deleted.