Batch of unorganized changes
Instead of Render going through Entities and figuring out what they
should look like, it is passed to each Entity which renders whatever the
hell it wants. This is a much more flexible system. Now a new Entity can
be added without adding compatibility for it into Render.

devmode button (F3) added

World has a Rectangle for bounds instead of containing the height and
width. This allows ingame objects to reference the rectangle instead of
referencing the world.

Started work on a Text object that should be able to render basic text.
It's not finished, so it won't actually work yet.