Add DevInfo, refactor Keyboard, comment
Game provides DevInfo with data about the game's status every second,
and if devmode is on, devmode's render method is called and it displays
the text on the screen. This allows easy addition of new data without
taking space in the titlebar, plus it looks cooler and paves the way for
future text display stuff.

Game.setup() has been added, allowing the Game to get its resources
prepared before it starts running. Just doing it in the constructor
doesn't really work out well because the buffer strategy can't be set
there for not-quite-understood reasons.

The Renderable interface seems to be taking some of the place of the
Entity class, so Entity's definition is being narrowed down to mean any
object in the game world. I like strong class definitions, and so should
you.

Text was deleted when I found out I didn't need it at all.

Render no longer contains a method to render a list of entities because
that's really none of its business. It can just keep rendering one thing
at a time and be happy it doesn't have to think any more than that.