Improved Keyboard memory usage
each control in keyboard is an array of bytes instead of shorts. AFAIK
it still matches all the keycode values provided by the KeyEvent class,
so it should be fine. I'm not sure if it matters whether the bytes are
signed or not. I hope it doesn't, will have to do proper testing at some
point.