Added Vector2D class
Vector2D is a representation of a vector. It uses the cartesian
horizontal and vertical length to store its data because that is easier
to use for movement and seems a better way of storing the data. It can
be accessed using both concepts of a vector however, potentially
allowing for polar coordinate systems on higher level ingame objects in
the future. I'm not sure if the setMagnitude method is the fastest way
of doing what it does, but at this point I don't even care. The other
methods should all be somewhat fast, I think. If I was thorough I would
do tests, but I'm lazy.