Added rendering hints, secured Game dimensions
Currently implemented rendering hints tell Graphics objects to favor
quality over speed. Somehow this has lead to an increase in fps on my
system, but perhaps not on all systems. I will add options in the future
to customize these settings.

Game.width and Game.height have been made private so not everyone can
use them all the time. That's a little too important of a resource to
make public, and the static thing could even become a problem when it
comes time to make the Artillery Toolkit.