Tweaked vectors, secured mover coordinates
Mover coordinates have been made private instead of protected. In the
future, this might make it easier if I ever have to make a transition to
using BigDecimal coordinates. Either way, it's better to have things
enclosed.

the isometric ratio was changed to 1 in preparation for Project Spiff.
I'll probably get rid of it altogether since I doubt this engine will
even be used for isometric stuff at this point.