Fullscreen support
Fullscreen is toggleable with the F12 key (customizable, of course).

I moved the JFrame code to Game, so now it can be changed any time.

Game has a new aspectRatio variable instead of a height variable, and
height is calculated by multiplying width by that. I chose that to make
different screen resolutions easier.