Global Transform Calculations, and pausing!
A transform can be applied to a vector. This should be useful for
calculating collision detection later on.

The code to scale and rotate an image is working (I think!!) I haven't
tested it thoroughly but I can't think of any other ways to break it
that I haven't tried yet.

Added the ability to pause/unpause the game by pressing escape.

Fixed a problem with Err, now it prints exception details correctly.