Camera tracking
A Camera holds an AffineTransform that Render uses to move images to the
correct spot before rendering. Sprites can't be transformed yet, but I
plan to implement a position transform in the future.

The camera has 3 booleans to control which elements of its Transform are
applied to the AffineTransform.

I also refactored some image class names.