Changes to save system and image rendering
The save system isn't ready for use yet, and won't be until the future
large overhauls of the entity and image system. I just changed some
stuff around so that the code compiles and isn't quite as confusing (I
hope).

The image rendering system is being redesigned to be more efficient.
There is a SpriteSheet that is supposed to hold all the images, while
entities hold onto what are essentially pointers with extra data like
position and a visibility flag.