Rendering redesign - rotation support
Adds support for image and camera rotation using the new rendering
pipeline. This update uses a Transform to represent the camera, so
future updates could allow for any Transform to function as a camera.
The only thing that will make it a camera is my say-so, along with any
camera-specific motion scripting. Scaling is next on the list, and then
I can move on to some physics.

Also updated the Keyboard class to be able to support the entire
keyboard. The keys array now holds as many values as there are keys.