Adding Behaviors
A behavior is a specific action that occurs over many entities at once.
Instead of putting the action in tick methods among all the other things
that get stuck in there, I used a Behavior class. This allows Components
to be free of action-related parts, and allows an action to happen over
every entity without fear of being interrupted by other actions changing
variables around. A prime example is for physics. First the gravity
Behavior will analyze the entities and apply gravity to them, then any
other similar forces will act, and then at the end of the tick cycle the
collision Behavior will analyze all the entities and deal with
collisions.

If multiple Behaviors do not affect entities in the same way, this could
even allow for multithreading.

Also, I might want to rename Behaviors to Systems.