Added textures and function controllers
A texture is an image that is repeated to fill an area. Right now only
AABBs are supported for texturing. There's an offset vector that changes
the location of the texture on the object it's being drawn on.  There's
some kind of problem with alpha values but I'll solve that later.

Function controllers are used to perform an action when a button is
pressed, and wait before doing it again.